# Conflicts:
#	.env.example
#	docker-compose.yml
#	src/main/java/com/warehouse/kafka/consumer/LowStockAlertConsumer.java
#	src/main/resources/application.yml
#	src/test/java/com/warehouse/kafka/KafkaStockAlertProducerRetryTest.java

+++ src/main/java/com/warehouse/dto/response/report/LowStockItem.java
@@ -5,7 +5,7 @@ public record LowStockItem(
         String sku,
         String name,
         String category,
-        int currentStock,
+        long currentStock,
         int minStock,
-        int deficit
+        long deficit
 ) {}
